+++ lib/src/runner/browser/iframe_test.dart
@@ -43,6 +43,8 @@ class IframeTest implements Test {
               new State(
                   new Status.parse(message['status']),
                   new Result.parse(message['result'])));
+        } else if (message['type'] == 'print') {
+          controller.print(message['line']);
         } else {
           assert(message['type'] == 'complete');
           controller.completer.complete();